ENH: Add fake#get_keynames(), useful e.g. for writing custom completions
As s:fake_codes and s:fake_cache are script-local, we cannot access them from outside, but the information therein is necessary to obtain a list of all possible fake data sources, e.g. when writing a custom completion function for a :command that takes a fake data source.
Add fake#get_keynames(), with an implementation that somewhat mirrors fake#has_keyname().